Finishing the character builder 'graphical' interface

+++ HeroesAndCreatureUnitTesting/Consts/CharacterConsts.cs
@@ -1,9 +1,9 @@
 ﻿namespace UnitTests.Consts {
     internal static class CharacterConsts {
-        public static readonly double DefaultHealth = 100.0;
+        public static readonly float DefaultHealth = 100f;
         public static readonly int DefaultAgility = 10;
         public static readonly int DefaultInitiative = 0;
-        public static readonly double DefaultPower = 1.0;
+        public static readonly float DefaultPower = 1f;
         public static readonly int DefaultPlace = -1;
     }
 }
