+++ HeroesAndCreatures/Characters/Abstraction/Human.cs
@@ -1,5 +1,5 @@
 ﻿namespace HeroesAndCreatures.Characters {
     public class Human : AbstractCharacter {
-        public Human(double health, int agility) : base(health, agility) { }
+        public Human(float health, int agility) : base(health, agility) { }
     }
 }
